fix(netplay): room code v3 — 32-bit nonce + correct ISO 7064 check digit [S4-review HIGH-1a, MEDIUM-1]
Two independent S4 review findings that both land in the room-code
codec, fixed together because each one changes the wire format and
shipping them separately would burn two breaking changes.

MEDIUM-1 — the check digit did NOT detect all single-character typos.
room_code.c computed the ISO 7064 hybrid recurrence with the
intermediate sum reduced mod 37 (M+1) instead of mod 36 (M). That let
the running product reach 0 as well as 36, and (37 - product) % 36 maps
BOTH to check char '1' — 37 product states collapsed onto 36 outputs, so
the documented "every single-character substitution is detected"
guarantee (room_code.h) was false. Measured against the shipped v2
codec: 2508 of 1,960,000 substitutions (0.128%) decoded silently to a
DIFFERENT endpoint. Worked counterexample, both ROOM_CODE_OK:
  "248BVXBAA4DNM1" -> 34.23.190.173
  "2E8BVXBAA4DNM1" -> 114.23.190.173
i.e. a one-character typo aimed ~15 s of UDP punch traffic at an
uninvolved third party and then reported HOST_OFFLINE, never "Invalid
room code". Fixed by solving the check character correctly (sum mod M),
whose product range is [1,36] and whose check map is a bijection.

HIGH-1a — the punch gate was only 12 bits deep. The S4a punch token is
SHA-256(domain || ip[4] || port_be[2] || nonce_be[])[0..7]; an attacker
who knows the advertised ip:port knows every input except the nonce. At
ROOM_CODE_NONCE_BITS 12 that was a 4096-value search against a host that
is a perfect oracle, drainable at ~60 datagrams/s => whole space in
<=68 s, and under UPnP the advertised port is stable so a past opponent
could do it months later. Widened to 32 bits (>=2.2 years at the same
rate). The entropy MUST live in the code: the joiner derives the token
from the decoded code alone on the direct-punch attempt, before any
rendezvous exchange, and the session key is what INDEXES the rendezvous
slot so it cannot be delivered by the thing it indexes. 5 bits/char
means 32 nonce bits cost exactly 4 more payload chars (48 fixed + 32 =
80 = 16*5) — the minimum possible.

Format v3 (room_code.h:68-110, room_code.c):
  '3' + 80-bit ip(32)|port(16)|nonce(32) in 16 Crockford chars + check
  = 18 chars, displayed XXXXXX-XXXXXX-XXXXXX (20 visible).
Cost, stated plainly: 4 more OSK presses and one more dash group.

- The 80-bit payload is wider than a uint64_t, so it is now carried as
  the SAME canonical 10-byte big-endian serialization the derivations
  hash (room_code_pack_payload / rendezvous.c REND_KEY_PAYLOAD_LEN),
  with the 16 base-32 chars cut out of that bitstream
  (room_code_get5/put5). One canonical byte order for the code and the
  hash input.
- Derivation domains bumped "3SXR-SK2"/"3SXR-PT2" -> "...SK3"/"...PT3"
  so a v2 payload can never land in a v3 rendezvous slot or produce a
  matching token (rendezvous.c). rendezvous.c also dropped the
  hardcoded `nonce > 0x0FFFu` reject and the 12-bit-into-2-byte
  serialization, both of which would have silently truncated.
- The '3SXR' WIRE version is deliberately UNCHANGED (still 2): the
  session key is a 16-byte opaque blob to the server, so a room-code
  format bump needs no server change.
- The defective recurrence is retained as iso7064_legacy_compute for
  ONE purpose — recognizing v1 (11-char) and v2 (14-char) codes so they
  surface as ROOM_CODE_OLD_FORMAT ("code is from an older version")
  instead of generic garbage. Checksum-invalid 11/14-char input stays
  MALFORMED: garbage must not masquerade as a version mismatch.
- Wrapper OSD kept in lockstep: DP2P_CODE_CHAR_LEN 14 -> 18, dashes at
  6/12, g_dp2p_code_buf and the Recently-Joined arrays sized from the
  new RECENT_JOIN_CODE_BUF (24), recent-joins record widened
  (kRecentCodeBuf, "%23s" scan, " %-18s %-9s" row = 29 of 32 columns).
  An 18-char code would have OVERFLOWED the old 16-byte buffers.

Tests — test_room_code.c gains full_alphabet_sweep(), the MEDIUM-1
regression net the old test could not be: the previous typo_detection
applied ONE hardcoded substitution per position (+1 in the alphabet), so
it never saw the collapse. The sweep replaces EVERY position with EVERY
other character of the full 0-9A-Z alphabet and asserts none decodes to
a different endpoint.

  post-fix: 0 undetected of 2,520,000 across 4000 codes, exit 0
  pre-fix (recurrence neutralized back to mod 37, everything else v3):
            3484 of 2,520,000 (0.1383%) undetected, exit 1

Also added: 32-bit-width assertions (ROOM_CODE_NONCE_BITS >= 32; nonces
differing only above bit 11 must produce different codes/keys/tokens and
round-trip exactly; CSPRNG draws must actually populate the high bits),
v2-OLD_FORMAT recognition, the MEDIUM-1 counterexample pair as a
standing case, and length-parametric bogus-length rejection.

Verified: host Debug (-DENABLE_NETPLAY_TESTS -DNETPLAY_TEST_HOOKS=ON),
all six netplay harnesses exit 0.

+++ src/netplay/rendezvous.h
@@ -12,11 +12,17 @@
  * (c) parse a 32-byte DELIVER or CHALLENGE, and (d) parse the
  * configured `udp://host:port` signal URL.
  *
- * Security note (v2, S4b): the session key is
- * SHA-256("3SXR-SK2" || ip[4] || port_be[2] || nonce_be[2])[0..15] —
- * the byte-aligned serialization of the v2 room-code payload including
- * the 12-bit CSPRNG nonce, domain-separated from the S4a punch token
- * ("3SXR-PT2"). It is derived (not transmitted) and not user-typeable.
+ * Security note (v3): the session key is
+ * SHA-256("3SXR-SK3" || ip[4] || port_be[2] || nonce_be[4])[0..15] —
+ * the byte-aligned serialization of the v3 room-code payload including
+ * the 32-bit CSPRNG nonce, domain-separated from the S4a punch token
+ * ("3SXR-PT3"). It is derived (not transmitted) and not user-typeable.
+ *
+ * The domain strings carry the room-code format version ("...2" -> "...3")
+ * so a build that somehow paired a v2 payload with a v3 derivation
+ * cannot land in the same rendezvous slot or produce a matching punch
+ * token. The '3SXR' WIRE version is unaffected and stays 2 — the
+ * session key is a 16-byte opaque blob to the server.
  */
 
 #ifndef NETPLAY_RENDEZVOUS_H
@@ -31,24 +37,25 @@ extern "C" {
 
 /*
  * Derive the 16-byte rendezvous session key from the public endpoint
- * tuple + nonce that both peers share via the room code (v2, S4b).
+ * tuple + nonce that both peers share via the room code (v3).
  *
  * `ip_be` follows the same convention as room_code.c: the uint32_t's
  * in-memory bytes are the four IPv4 octets in network byte order (i.e.
  * the value returned by inet_pton(AF_INET, ...) into `s_addr`).
- * `nonce` is the 12-bit room-code nonce (host order, <= 0x0FFF).
+ * `nonce` is the 32-bit room-code nonce (host order; every value is
+ * in range).
  *
- * The derivation is domain-separated and covers the full v2 payload:
+ * The derivation is domain-separated and covers the full v3 payload:
  *
- *   key = SHA-256("3SXR-SK2" || ip[4] || port_be[2] || nonce_be[2])[0..15]
+ *   key = SHA-256("3SXR-SK3" || ip[4] || port_be[2] || nonce_be[4])[0..15]
  *
- * where nonce_be is the nonce as a 16-bit big-endian value (top 4 bits
- * zero). Because the 12-bit CSPRNG nonce is inside the hash, an
- * attacker who merely knows/guesses (ip, port) can no longer derive
- * the session key to squat or race the rendezvous slot (S4b) — only
- * someone holding the actual room code can. BREAKING change from the
- * v1 derivation (SHA-256 over the bare 6-byte payload), shipped
- * together with the v2 room-code format to the whole alpha group.
+ * where nonce_be is the nonce as a 32-bit big-endian value. Because the
+ * 32-bit CSPRNG nonce is inside the hash, an attacker who merely
+ * knows/guesses (ip, port) can no longer derive the session key to
+ * squat or race the rendezvous slot — only someone holding the actual
+ * room code can. BREAKING change from the v2 derivation (12-bit nonce,
+ * "3SXR-SK2"), shipped together with the v3 room-code format to the
+ * whole alpha group.
  *
  * Writes 16 bytes into `out_key`. Returns true on success. On failure
  * (ip_be == 0, indicating an unset / invalid endpoint), zeroes
@@ -56,7 +63,7 @@ extern "C" {
  */
 bool Rendezvous_DeriveSessionKey(uint32_t ip_be,
                                  uint16_t public_port,
-                                 uint16_t nonce,
+                                 uint32_t nonce,
                                  uint8_t out_key[16]);
 
 /*
@@ -65,7 +72,7 @@ bool Rendezvous_DeriveSessionKey(uint32_t ip_be,
  * payload. Same inputs and payload serialization as
  * Rendezvous_DeriveSessionKey, different domain string:
  *
- *   token = SHA-256("3SXR-PT2" || ip[4] || port_be[2] || nonce_be[2])[0..7]
+ *   token = SHA-256("3SXR-PT3" || ip[4] || port_be[2] || nonce_be[4])[0..7]
  *
  * so the two derivations can never collide. The token proves knowledge
  * of the room-code payload INCLUDING the nonce: the host only accepts
@@ -75,13 +82,18 @@ bool Rendezvous_DeriveSessionKey(uint32_t ip_be,
  * course derive it — the room code IS the shared secret in this
  * friend-to-friend model. S5's relay will reuse this token.
  *
+ * The nonce is 32 bits as of v3. At 12 bits the token had only 4096
+ * unguessable variants against a host that answered every guess (accept
+ * -> handoff, miss -> silent drop, no cap), which the host receive path
+ * could be walked through in under 68 seconds. See room_code.h.
+ *
  * Writes 8 bytes into out_token. Returns true on success; on failure
  * (ip_be == 0 or hash failure) zeroes out_token and returns false.
  */
 #define REND_PUNCH_TOKEN_LEN 8
 bool Rendezvous_DerivePunchToken(uint32_t ip_be,
                                  uint16_t public_port,
-                                 uint16_t nonce,
+                                 uint32_t nonce,
                                  uint8_t out_token[REND_PUNCH_TOKEN_LEN]);
 
 /* S4c protocol v2 sizes: REGISTER/POLL carry an 8-byte return-
